Fixing bug on FileService
The error message "Text 'Date' could not be parsed at index 0" is thrown when the LocalDateTime.parse() method is unable to parse the provided text into a date time value.

It appears that the code is trying to parse the string "Date" instead of the actual date string that is supposed to be parsed.
It seems that there is an error in the code that is causing the wrong string to be passed to the LocalDateTime.parse() method.

You should check the parsedDate variable, it should contains a string that can be parsed by the DateTimeFormatter.ofPattern("MMM-yy")
You can do this by adding a debug print statement to print the parsedDate before calling the LocalDateTime.parse method.

It could be that the split method is not working as expected, it should split the line by "," and assign the first element of the resulting array to parsedDate.
Also, it's important to check the data format in the csv file, it should match the pattern "MMM-yy"

If the parsedDate variable is correct, you should check the format of the date time string and make sure that it matches the pattern specified in the DateTimeFormatter.ofPattern("MMM-yy")

If the problem persist, you can also add a catch block for the DateTimeParseException and print the exception message to understand where the problem is coming from.